refactor: load credential files only in node
